requirements.txt
aioredis==2.0.1
async-timeout==5.0.1
python-dotenv==1.0.1
redis==5.2.1
tornado==6.4.2
typing_extensions==4.12.2
